github-bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.0
PyGithub>=2.1.1
